LocationGroup -> ts
